Power class now has ability timer that tracks cooldown.
Player still responsible for the ability itself but
whether the ability can be used is tracked by Power.

Also added a duration on the power that prevents collection
of more power while it's running. This may be changed later,
in which case the Power timer needs to be refactored to
test for curValue == 0.